refactor(server): Move AbstractRuleSet to zally-core #560
