+++ server/src/main/java/de/zalando/zally/rule/zalando/ZalandoRuleSet.kt
@@ -1,6 +1,6 @@
 package de.zalando.zally.rule.zalando
 
-import de.zalando.zally.rule.AbstractRuleSet
+import de.zalando.zally.core.AbstractRuleSet
 import de.zalando.zally.rule.api.Rule
 import java.net.URI
