+++ server/src/test/java/de/zalando/zally/apireview/RestApiTestConfiguration.kt
@@ -5,7 +5,9 @@ import com.typesafe.config.Config
 import de.zalando.zally.core.EMPTY_JSON_POINTER
 import de.zalando.zally.core.toJsonPointer
 import de.zalando.zally.rule.RulesManager
-import de.zalando.zally.rule.AbstractRuleSet
+import de.zalando.zally.core.AbstractRuleSet
+import de.zalando.zally.core.EMPTY_JSON_POINTER
+import de.zalando.zally.core.toJsonPointer
 import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
