+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/AbstractRuleSet.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import de.zalando.zally.rule.api.RuleSet
